chore: seperate preset and location view
