+++ src/frontend/sharedComponents/Editor/index.tsx
@@ -2,8 +2,12 @@ import * as React from 'react';
 import {Photo} from '../../contexts/PhotoPromiseContext/types';
 import {DescriptionField} from './DescriptionField';
 import {MediaScrollView} from '../MediaScrollView';
-import {PresetAndLocationView} from './PresetAndLocationView';
 import {ScreenContentWithDock} from '../ScreenContentWithDock';
+import {StyleSheet, View} from 'react-native';
+import {LIGHT_GREY} from '../../lib/styles';
+import {PresetView} from './PresetView';
+import {LocationView} from './LocationView';
+import {Divider} from '../Divider';
 
 type EditorProps = {
   presetName: string;
@@ -21,27 +25,36 @@ type EditorProps = {
 };
 
 export const Editor = ({
-  presetName,
-  onPressPreset,
-  PresetIcon,
   notes,
   updateNotes,
   photos,
   location,
   actionsRow,
+  ...presetProps
 }: EditorProps) => {
   return (
     <ScreenContentWithDock
       dockContainerStyle={{padding: 0}}
       dockContent={actionsRow}>
-      <PresetAndLocationView
-        onPressPreset={onPressPreset}
-        presetName={presetName}
-        PresetIcon={PresetIcon}
-        location={location}
-      />
+      <View style={styles.container}>
+        <PresetView {...presetProps} />
+        {location && (
+          <>
+            <Divider />
+            <LocationView {...location} />
+          </>
+        )}
+      </View>
       <DescriptionField notes={notes} updateNotes={updateNotes} />
       <MediaScrollView photos={photos} />
     </ScreenContentWithDock>
   );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    borderRadius: 12,
+    borderWidth: 1,
+    borderColor: LIGHT_GREY,
+  },
+});
